move resources to testclasses in reference
