+++ src/test/java/spoon/test/reference/testclasses/MyClass.java
@@ -1,4 +1,4 @@
-package spoon.test.reference;
+package spoon.test.reference.testclasses;
 
 public class MyClass {
 	public MyClass() {
